Beautify the BUILD_TIMESTAMP and put it into the binary.
* configure.ac (BUILD_TIMESTAMP): Use an ISO date string on all
platforms and ac_define it.
* src/sysutils.c (_assuan_sysutils_blurb): But revision and build date
into the binary.
--

An ISO date is for sure better than the old C standard time format.
An almost all platforms this should work.  We don't print the seconds
to be compatible with the way we did in on Windows.  However we print
in UTC.