Add hack to have different names for 64 bit Windows DLLs.
* ltmain.sh: Prefix the SO number for W64 with a "6".
--

We need to stick to libtool 2.4.2 anyway, thus we take the easy way
and hack libtool instead of adding "-release 6" to the Makefile.  This
also avoids an extra dash before the 6 and thus yields a prettier name
than the one produced by the -release option.  The header files and
def files are supposed to be identical for w32 and w64 thus we don't
need to do any special code for them.

Having a different namespace for 64 bit DLLs is such a good idea that
I hope this or a similar feature will be provided by future libtool
versions.