Show the pid of listening process in the hello line.
* src/assuan-listen.c (assuan_accept): Use getpid instead of
assuan_get_pid.

--

This changes the initial interaction.  Before the change, the PID of
peer was shown from listening process (but not used at all, in the
code).  That is, it was the PID of connecting process.  After the
change, the PID of listening process is shown, which can be used by
connecting process.

GnuPG-bug-id: 6236
Fixes-commit: a00b1ede5403f1e38534cc3b232739f20b8567e9
Signed-off-by: NIIBE Yutaka <gniibe@fsij.org>